+++ packages/form-render/src/useForm.js
@@ -35,6 +35,7 @@ export const useForm = props => {
   const clickSubmit = useRef(false); // 点击submit的那一下，不要执行useEffect里的validate
   const beforeFinishRef = useRef();
   const localeRef = useRef('cn');
+  const validateMessagesRef = useRef();
   const _data = useRef({}); // 用ref是为了破除闭包的影响
   const _touchedKeys = useRef([]); // 用ref是为了破除闭包的影响
 
@@ -98,6 +99,7 @@ export const useForm = props => {
           isRequired: true,
           touchedKeys: _touchedKeys.current,
           locale: localeRef.current,
+          validateMessages: validateMessagesRef.current,
         }).then(res => {
           const oldFormatErrors = res.map(item => item.name);
           _onValidate(oldFormatErrors);
@@ -118,6 +120,7 @@ export const useForm = props => {
       isRequired: allTouched,
       touchedKeys: _touchedKeys.current,
       locale: localeRef.current,
+      validateMessages: validateMessagesRef.current,
     }).then(res => {
       _setErrors(res);
     });
@@ -142,11 +145,18 @@ export const useForm = props => {
   //   { name: 'a.b.c', errors: ['Please input your Password!', 'something else is wrong'] },
   // ]
 
-  const syncStuff = ({ schema, flatten, beforeFinish, locale }) => {
+  const syncStuff = ({
+    schema,
+    flatten,
+    beforeFinish,
+    locale,
+    validateMessages,
+  }) => {
     schemaRef.current = schema;
     flattenRef.current = flatten;
     beforeFinishRef.current = beforeFinish;
     localeRef.current = locale;
+    validateMessagesRef.current = validateMessages;
   };
 
   // TODO: 外部校验的error要和本地的合并么？
@@ -193,6 +203,7 @@ export const useForm = props => {
       touchedKeys: [],
       isRequired: true,
       locale: localeRef.current,
+      validateMessages: validateMessagesRef.current,
     })
       .then(errors => {
         // 如果有错误，也不停止校验和提交，在onFinish里让用户自己搞
@@ -233,6 +244,7 @@ export const useForm = props => {
 
   const resetFields = () => {
     _setData({});
+    _setErrors([]);
   };
 
   // const setValue = (id, value, dataIndex) => {