+++ tools/schema-generator/src/Right/ItemSettings.js
@@ -1,7 +1,7 @@
-import React from 'react';
-import FRWrapper from '../FRWrapper';
+import React, { useState, useEffect } from 'react';
+import FormRender, { useForm } from 'form-render';
 import { useStore } from '../hooks';
-import { widgets } from '../widgets/antd';
+import { widgets as defaultWidgets } from 'form-render/src/widgets/antd';
 import IdInput from '../widgets/antd/idInput';
 import PercentSlider from '../widgets/antd/percentSlider';
 import {
@@ -11,14 +11,19 @@ import {
   advancedElements,
   layouts,
 } from '../Settings';
-import { getWidgetName } from '../mapping';
+import { getWidgetName } from 'form-render/src/mapping';
 import { isObject } from '../utils';
 
 export default function ItemSettings() {
+  const form = useForm();
   const { selected, flatten, onItemChange, userProps = {} } = useStore();
   const { settings, commonSettings } = userProps;
-  let settingSchema = {};
-  let settingData = {};
+  const [settingSchema, setSettingSchema] = useState({});
+  const widgets = {
+    ...defaultWidgets,
+    idInput: IdInput,
+    percentSlider: PercentSlider,
+  };
 
   const getWidgetList = (settings, commonSettings) => {
     let widgetList = [];
@@ -33,12 +38,12 @@ export default function ItemSettings() {
     return widgetList;
   };
 
-  const onDataChange = newSchema => {
-    if (selected) {
+  const onDataChange = (value) => {
+    if (selected && value.$id) {
       try {
         const item = flatten[selected];
         if (item && item.schema) {
-          onItemChange(selected, { ...item, schema: newSchema });
+          onItemChange(selected, { ...item, schema: value });
         }
       } catch (error) {
         console.log(error, 'catch');
@@ -46,58 +51,54 @@ export default function ItemSettings() {
     }
   };
 
-  // 算widgetList
-  const _settings = Array.isArray(settings)
-    ? [...settings, { widgets: [...elements, ...advancedElements, ...layouts] }] // TODO: 不是最优解
-    : defaultSettings;
-  const _commonSettings = isObject(commonSettings)
-    ? commonSettings
-    : defaultCommonSettings;
-  const widgetList = getWidgetList(_settings, _commonSettings);
+  useEffect(() => {
+    // 算widgetList
+    const _settings = Array.isArray(settings)
+      ? [...settings, { widgets: [...elements, ...advancedElements, ...layouts] }] // TODO: 不是最优解
+      : defaultSettings;
+    const _commonSettings = isObject(commonSettings)
+      ? commonSettings
+      : defaultCommonSettings;
+    const widgetList = getWidgetList(_settings, _commonSettings);
 
-  // setting该显示什么的计算，要把选中组件的schema和它对应的widgets的整体schema进行拼接
-  let itemSelected;
-  let widgetName;
-  try {
-    itemSelected = flatten[selected];
-    if (itemSelected) {
-      widgetName = getWidgetName(itemSelected.schema);
-    }
-    if (widgetName) {
-      // const name = getKeyFromUniqueId(selected);
-      const element = widgetList.find(e => e.widget === widgetName) || {}; // 有可能会没有找到
-      const schemaNow = element.setting;
-      settingSchema = {
-        schema: {
+    // setting该显示什么的计算，要把选中组件的schema和它对应的widgets的整体schema进行拼接
+    let itemSelected;
+    let widgetName;
+    try {
+      itemSelected = flatten[selected];
+      if (itemSelected) {
+        widgetName = getWidgetName(itemSelected.schema);
+      }
+      if (widgetName) {
+        // const name = getKeyFromUniqueId(selected);
+        const element = widgetList.find(e => e.widget === widgetName) || {}; // 有可能会没有找到
+        const schemaNow = element.setting;
+        setSettingSchema({
           type: 'object',
+          displayType: 'column',
+          showDescIcon: false,
           properties: {
             ...schemaNow,
           },
-        },
-      };
-      settingData = itemSelected.schema;
+        });
+        setTimeout(() => {
+          form.setValues(itemSelected.schema);
+        })
+      }
+    } catch (error) {
+      console.log(error);
     }
-  } catch (error) {
-    console.log(error);
-  }
-
-  const _widgets = {
-    ...widgets,
-    idInput: IdInput,
-    percentSlider: PercentSlider,
-  };
-
-  // TODO2: 这边开放
+  }, [selected]);
 
   return (
     <div style={{ paddingRight: 24 }}>
-      <FRWrapper
+      <FormRender
+        form={form}
         schema={settingSchema}
-        formData={settingData}
-        onChange={onDataChange}
-        widgets={_widgets}
-        preview={true}
-        frProps={{ displayType: 'column', showDescIcon: false }}
+        widgets={widgets}
+        watch={{
+          '#': v => onDataChange(v)
+        }}
       />
     </div>
   );